Use double instead of float to (avoid) numerical computation problems It will be better to use GNU MPFR Library

+++ src/lib/3rdparty/csgjs/csgjs.cpp
@@ -1,9 +1,28 @@
+// Original CSG.JS library by Evan Wallace (http://madebyevan.com), under the MIT license.
+// GitHub: https://github.com/evanw/csg.js/
+//
+// C++ port by Tomasz Dabrowski (http://28byteslater.com), under the MIT license.
+// GitHub: https://github.com/dabroz/csgjs-cpp/
+//
+// Using double instead of float
+// Laurent Ribon (http://glc-lib.sourceforge.net)
+// GitHub: https://github.com/laumaya/GLC_lib/
+//
+// Constructive Solid Geometry (CSG) is a modeling technique that uses Boolean
+// operations like union and intersection to combine 3D solids. This library
+// implements CSG operations on meshes elegantly and concisely using BSP trees,
+// and is meant to serve as an easily understandable implementation of the
+// algorithm. All edge cases involving overlapping coplanar polygons in both
+// solids are correctly handled.
+//
+// To use this as a header file, define CSGJS_HEADER_ONLY before including this file.
+//
 
 #include "csgjs.h"
 
 // `CSG.Plane.EPSILON` is the tolerance used by `splitPolygon()` to decide if a
 // point is on the plane.
-static const float csgjs_EPSILON = 0.00001f;
+static const float csgjs_EPSILON = 0.00001;
 
 struct csgjs_plane;
 struct csgjs_polygon;
@@ -13,7 +32,7 @@ struct csgjs_node;
 struct csgjs_plane
 {
     csgjs_vector normal;
-    float w;
+    double w;
 
     csgjs_plane();
     csgjs_plane(const csgjs_vector & a, const csgjs_vector & b, const csgjs_vector & c);
@@ -68,12 +87,12 @@ struct csgjs_csgnode
 
 inline static csgjs_vector operator + (const csgjs_vector & a, const csgjs_vector & b) { return csgjs_vector(a.x + b.x, a.y + b.y, a.z + b.z); }
 inline static csgjs_vector operator - (const csgjs_vector & a, const csgjs_vector & b) { return csgjs_vector(a.x - b.x, a.y - b.y, a.z - b.z); }
-inline static csgjs_vector operator * (const csgjs_vector & a, float b) { return csgjs_vector(a.x * b, a.y * b, a.z * b); }
-inline static csgjs_vector operator / (const csgjs_vector & a, float b) { return a * (1.0f / b); }
-inline static float dot(const csgjs_vector & a, const csgjs_vector & b) { return a.x * b.x + a.y * b.y + a.z * b.z; }
-inline static csgjs_vector lerp(const csgjs_vector & a, const csgjs_vector & b, float v) { return a + (b - a) * v; }
-inline static csgjs_vector negate(const csgjs_vector & a) { return a * -1.0f; }
-inline static float length(const csgjs_vector & a) { return sqrtf(dot(a, a)); }
+inline static csgjs_vector operator * (const csgjs_vector & a, double b) { return csgjs_vector(a.x * b, a.y * b, a.z * b); }
+inline static csgjs_vector operator / (const csgjs_vector & a, double b) { return a * (1.0 / b); }
+inline static double dot(const csgjs_vector & a, const csgjs_vector & b) { return a.x * b.x + a.y * b.y + a.z * b.z; }
+inline static csgjs_vector lerp(const csgjs_vector & a, const csgjs_vector & b, double v) { return a + (b - a) * v; }
+inline static csgjs_vector negate(const csgjs_vector & a) { return a * -1.0; }
+inline static double length(const csgjs_vector & a) { return sqrt(dot(a, a)); }
 inline static csgjs_vector unit(const csgjs_vector & a) { return a / length(a); }
 inline static csgjs_vector cross(const csgjs_vector & a, const csgjs_vector & b) { return csgjs_vector(a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x); }
 
@@ -90,7 +109,7 @@ inline static csgjs_vertex flip(csgjs_vertex v)
 // Create a new vertex between this vertex and `other` by linearly
 // interpolating all properties using a parameter of `t`. Subclasses should
 // override this to interpolate additional properties.
-inline static csgjs_vertex interpolate(const csgjs_vertex & a, const csgjs_vertex & b, float t)
+inline static csgjs_vertex interpolate(const csgjs_vertex & a, const csgjs_vertex & b, double t)
 {
     csgjs_vertex ret;
     ret.pos = lerp(a.pos, b.pos, t);
@@ -101,19 +120,19 @@ inline static csgjs_vertex interpolate(const csgjs_vertex & a, const csgjs_verte
 
 // Plane implementation
 
-csgjs_plane::csgjs_plane() : normal(), w(0.0f)
+csgjs_plane::csgjs_plane() : normal(), w(0.0)
 {
 }
 
 bool csgjs_plane::ok() const
 {
-    return length(this->normal) > 0.0f;
+    return length(this->normal) > 0.0;
 }
 
 void csgjs_plane::flip()
 {
     this->normal = negate(this->normal);
-    this->w *= -1.0f;
+    this->w *= -1.0;
 }
 
 csgjs_plane::csgjs_plane(const csgjs_vector & a, const csgjs_vector & b, const csgjs_vector & c)
@@ -144,7 +163,7 @@ void csgjs_plane::splitPolygon(const csgjs_polygon & polygon, std::vector<csgjs_
 
     for (size_t i = 0; i < polygon.vertices.size(); i++)
     {
-        float t = dot(this->normal, polygon.vertices[i].pos) - this->w;
+        double t = dot(this->normal, polygon.vertices[i].pos) - this->w;
         int type = (t < -csgjs_EPSILON) ? BACK : ((t > csgjs_EPSILON) ? FRONT : COPLANAR);
         polygonType |= type;
         types.push_back(type);
@@ -183,7 +202,7 @@ void csgjs_plane::splitPolygon(const csgjs_polygon & polygon, std::vector<csgjs_
                 if (ti != FRONT) b.push_back(vi);
                 if ((ti | tj) == SPANNING)
                 {
-                    float t = (this->w - dot(this->normal, vi.pos)) / dot(this->normal, vj.pos - vi.pos);
+                    double t = (this->w - dot(this->normal, vi.pos)) / dot(this->normal, vj.pos - vi.pos);
                     csgjs_vertex v = interpolate(vi, vj, t);
                     f.push_back(v);
                     b.push_back(v);
